Updated OrderDaoImplTest to test also negative cases

+++ persistence/src/main/java/cz/fi/muni/pa165/dao/OrderDaoImpl.java
@@ -4,6 +4,7 @@ import cz.fi.muni.pa165.entity.Order;
 import cz.fi.muni.pa165.entity.Service;
 import cz.fi.muni.pa165.entity.Tire;
 import cz.fi.muni.pa165.entity.User;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
@@ -23,6 +24,9 @@ public class OrderDaoImpl implements OrderDao {
 
     @Override
     public void create(Order order) {
+        if (order == null) {
+            throw new DataAccessException("Attempting to create null order."){};
+        }
         entityManager.persist(order);
     }
 
@@ -34,31 +38,48 @@ public class OrderDaoImpl implements OrderDao {
 
     @Override
     public Order findById(Long id) {
+        if (id == null) {
+            throw new DataAccessException("Cannot find order, because id is null."){};
+        }
         return entityManager.find(Order.class, id);
     }
 
     @Override
     public void remove(Order order) {
+        if (order == null) {
+            throw new DataAccessException("Attempting to remove null order."){};
+        }
         entityManager.remove(order);
     }
 
     @Override
     public void update (Order order) {
+        if (order == null) {
+            throw new DataAccessException("Attempting to update null order."){};
+        }
         entityManager.merge(order);
     }
 
     @Override
-    public List<Order> findByUser(User u) {
+    public List<Order> findByUser(User user) {
+        if (user == null) {
+            throw new DataAccessException("Attempting to find order by null user."){};
+        }
         TypedQuery<Order> query = entityManager.createQuery(
                 "Select o from Order o where o.user = :userid",
                 Order.class);
 
-        query.setParameter("userid", u);
+        query.setParameter("userid", user);
         return query.getResultList();
     }
 
     @Override
     public void addService(Long orderId, Long serviceID) {
+        if (orderId == null) {
+            throw new DataAccessException("Cannot add service to order, because orderId is null."){};
+        } else if (serviceID == null) {
+            throw new DataAccessException("Cannot add service to order, because serviceID is null."){};
+        }
         Order order = findById(orderId);
         Service service = entityManager.find(Service.class, serviceID);
         order.getServices().add(service);
@@ -67,6 +88,11 @@ public class OrderDaoImpl implements OrderDao {
 
     @Override
     public void addTire(Long orderId, Long tireID) {
+        if (orderId == null) {
+            throw new DataAccessException("Cannot add service to order, because orderId is null."){};
+        } else if (tireID == null) {
+            throw new DataAccessException("Cannot add service to order, because tireID is null."){};
+        }
         Order order = findById(orderId);
         Tire tire = entityManager.find(Tire.class, tireID);
         order.getTires().add(tire);
